Fixed comments
* https://github.com/TouchInstinct/RoboSwag/pull/69#discussion_r324111996 – renamed "edittextutils" to "edittextformatters"
* https://github.com/TouchInstinct/RoboSwag/pull/69#discussion_r324112354 – removed return statement
* https://github.com/TouchInstinct/RoboSwag/pull/69#discussion_r324112841 – removed "with"
