+++ edit-text-formatters/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest
+    package="ru.touchin.edittextformatters"/>
