+++ edit-text-formatters/src/main/java/ru/touchin/edittextformatters/converter/Extensions.kt
@@ -1,0 +1,11 @@
+package ru.touchin.edittextformatters.converter
+
+import java.math.BigDecimal
+
+fun String.toBigDecimalOrZeroWithoutGrouping(): BigDecimal = this.toBigDecimalOrNullWithoutGrouping() ?: BigDecimal.ZERO
+
+private fun String.toBigDecimalOrNullWithoutGrouping(): BigDecimal? = try {
+    BigDecimal(this.replace("\\s".toRegex(), ""))
+} catch (_: NumberFormatException) {
+    null
+}
