--- edit-text-utils/src/main/java/ru/touchin/edittextutils/converter/wrap/Convertible.kt
@@ -1,30 +1,0 @@
-package ru.touchin.edittextutils.converter.wrap
-
-import android.text.TextWatcher
-import android.text.method.KeyListener
-import androidx.annotation.ColorInt
-
-/**
- * Interface for Custom Views usage
- */
-interface Convertible {
-
-    fun setText(charSequence: CharSequence, placeCursorToTheEnd: Boolean = true, incrementCursorPosition: Boolean = false)
-
-    fun getText(): CharSequence
-
-    fun addTextChangedListener(watcher: TextWatcher)
-
-    fun setKeyListener(keyListener: KeyListener)
-
-    fun isEnabled(): Boolean
-
-    fun setEnabled(enabled: Boolean)
-
-    fun setTextColor(@ColorInt colorInt: Int)
-
-    fun addOnFocusChangedListener(onFocusChangedListener: (Boolean) -> Unit)
-
-    fun isFocused(): Boolean
-
-}
