--- edit-text-utils/src/main/java/ru/touchin/edittextutils/verifier/verifiers/DecimalVerifier.kt
@@ -1,20 +1,0 @@
-package ru.touchin.edittextutils.verifier.verifiers
-
-import ru.touchin.edittextutils.commands.Command
-import java.math.BigDecimal
-
-/**
- * Verifies decimal numbers in text input.
- * @param maxDecimalNumber maximum allowed numbers after separator.
- */
-class DecimalVerifier(maxDecimalNumber: Int) : Verifier<BigDecimal> {
-
-    private val pattern = "\\d+(\\.?)(\\d{0,$maxDecimalNumber})".toRegex()
-
-    override fun verify(text: String): Command<BigDecimal> = if (pattern.matches(text) == true) {
-        Command.Success()
-    } else {
-        Command.Fallback()
-    }
-
-}
