--- edit-text-utils/src/main/java/ru/touchin/edittextutils/verifier/verifiers/MaximumValueVerifier.kt
@@ -1,18 +1,0 @@
-package ru.touchin.edittextutils.verifier.verifiers
-
-import ru.touchin.edittextutils.commands.Command
-import java.math.BigDecimal
-
-/**
- * Verifies value of text input.
- * @param maxValue maximum allowed value in text input.
- */
-class MaximumValueVerifier(val maxValue: BigDecimal) : Verifier<BigDecimal> {
-
-    override fun verify(text: String): Command<BigDecimal> = if (text.toBigDecimal() <= maxValue) {
-        Command.Success()
-    } else {
-        Command.Set(maxValue)
-    }
-
-}
