+++ modules.gradle
@@ -5,7 +5,7 @@ if (gradle.ext.has('componentsRoot')) {
     rootDir = settingsDir
 }
 
-include ':edit-text-utils'
+include ':edit-text-formatters'
 include ':logging'
 include ':utils'
 include ':navigation'
@@ -18,7 +18,7 @@ include ':recyclerview-adapters'
 include ':kotlin-extensions'
 include ':templates'
 
-project(':edit-text-utils').projectDir = new File(rootDir, 'edit-text-utils')
+project(':edit-text-formatters').projectDir = new File(rootDir, 'edit-text-formatters')
 project(':utils').projectDir = new File(rootDir, 'utils')
 project(':logging').projectDir = new File(rootDir, 'logging')
 project(':navigation').projectDir = new File(rootDir, 'navigation')
